Add specific check for Timeout::Error and cause our own exception to be fired.
Causes issues because Timeout::Error is not a subclass of StandardError, it is a subclass of the Interrupt class.

http://lindsaar.net/2007/12/9/rbuf_filltimeout-error

This solves an unexpected exception when Armory hangs due to missing data, interestingly a Firefox user agent gets a hang, another user agent that causes HTML to be rended gets an immediate response.